requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-datatables-view==1.19.1
django-pg-returning==1.2.4
idna==2.8
importlib-metadata==1.4.0
infi.clickhouse-orm==1.2.0
iso8601==0.1.12
kombu==4.6.7
psycopg2==2.8.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
statsd==3.3.0
typing==3.7.4.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
zipp==2.1.0
